Fixed the database encoding issue by reusing the conversion function from the EmailParser_FSM.cpp. The database reader now works as expected, however, there is an issue where the conversion changes the file bytes which means it generates a different hash compared to the original email.
Also half fixed a bug where if a Mime Multipart email has lines inbetween the end of the email header and the first boundary, the FSM will read the lines inbetween as boundaries. I fixed this by adding a preamble state inbetween the end of the headers and the start of the first Mime part. This works for now, however, I would like to include the skipped lines when rendering the email.